lessons/2021-05-16_sqlalchemy_transactions/requirements.txt
cffi==1.14.5
cryptography==3.4.7
pycparser==2.20
PyMySQL==1.0.2
python-dotenv==0.17.1
six==1.16.0
SQLAlchemy==1.3.24
SQLAlchemy-Utils==0.37.2
